Update composer and refactor code under the phpstan guide

+++ tests/Feature/VoyagerDataExportViewTest.php
@@ -25,7 +25,7 @@ class VoyagerDataExportViewTest extends TestCase implements ICommandStatus
     {
 
         $this->artisan("voyager:data:transport:export-data:view {$this->_getTableName()}")
-            ->assertExitCode(self::ALL_PROCESS_SUCCESS_CODE);
+            ->assertExitCode( (int) self::ALL_PROCESS_SUCCESS_CODE );
     }
 
     /**
