+++ tests/Feature/VoyagerDataTransportExportControllerTest.php
@@ -33,7 +33,7 @@ class VoyagerDataTransportExportControllerTest extends TestCase implements IComm
     public function test_command(): void
     {
         $this->artisan("voyager:data:export:controller {$this->_getTableName()}")
-            ->assertExitCode(self::ALL_PROCESS_SUCCESS_CODE);
+            ->assertExitCode( (int) self::ALL_PROCESS_SUCCESS_CODE );
     }
 
     /**
