+++ tests/Feature/VoyagerDataTransportImportPermissionTest.php
@@ -24,6 +24,6 @@ class VoyagerDataTransportImportPermissionTest extends TestCase implements IComm
     public function test_command(): void
     {
         $this->artisan("voyager:data:transport:import:permission {$this->_getTableName()}")
-            ->assertExitCode(self::ALL_PROCESS_SUCCESS_CODE);
+            ->assertExitCode( (int) self::ALL_PROCESS_SUCCESS_CODE );
     }
 }
